+++ test/unit/gen_class/yum_get_release.py
@@ -16,8 +16,9 @@
 # Standard
 import sys
 import os
+import distro
+import socket
 import unittest
-import mock
 
 # Local
 sys.path.append(os.getcwd())
@@ -53,15 +54,12 @@ class UnitTest(unittest.TestCase):
             print("Error: Python 3 does not support yum==3.4.3, skipping test")
             self.skipTest("Pre-conditions not met.")
 
-        self.host_name = "HOSTNAME"
-        self.osys = "Linux"
-        self.release = "2.6"
-        self.distro = ("Centos", "7.5.1804", "Core")
+        self.host_name = socket.gethostname()
+        self.osys = distro.name()
+        self.release = distro.version()
+        self.distro = (distro.name(), distro.version(), distro.codename())
 
-    @mock.patch("platform.linux_distribution")
-    @mock.patch("platform.release")
-    @mock.patch("platform.system")
-    def test_get_release(self, mock_system, mock_release, mock_distro):
+    def test_get_release(self):
 
         """Function:  test_get_release
 
@@ -70,10 +68,6 @@ class UnitTest(unittest.TestCase):
         Arguments:
 
         """
-
-        mock_system.return_value = "Linux"
-        mock_release.return_value = "2.6"
-        mock_distro.return_value = ("Centos", "7.5.1804", "Core")
 
         yum = gen_class.Yum(self.host_name)