+++ CHANGELOG.md
@@ -21,10 +21,12 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - gen_class.Dnf.fetch_update_pkgs: Returns list of update packages - backwards comptable for gen_class.Yum.
 
 ### Changed
+- gen_class.Yum.\_\_init\_\_: Replaced platform module with distro module.
 - gen_class.Dnf.\_\_init\_\_: Replaced platform module with distro module.
 - Documentation changes.
 
 ### Removed
+- Removed platform module.
 - gen_class.KeyCaseInsensitiveDict.has_key: Dictionarys has_key has been removed in Python 3.